essential-requirements.txt
torch
torchvision
torchaudio
transformers
numpy
pillow
opencv-python-headless
pandas
qdrant-client
pydantic
streamlit 
num2words
accelerate>=0.26.0
openai
google-generativeai
google-genai
torchaudio
ffmpeg-python
pyannote-audio
pyyaml